+++ packages/@steedos-widgets/amis-lib/src/lib/converter/amis/toolbars/edit/copyListview.js
@@ -1,4 +1,5 @@
 export const copyListview = ()=>{
+  
     return {
         "type": "button",
         "label": "复制",
@@ -28,11 +29,26 @@ export const copyListview = ()=>{
                       "defaultData": {
                         "&": "${list_view}",
                         "name":"",
-                        "label": "${list_view.label} 的副本",
+                        "label": "${list_view.label}的副本",
                         "shared":false
                       },
-                      "fieldsExtend": fieldsExtend(),
-                      "fields": fields(),
+                      "fieldsExtend": "{\n  \"label\": {\n    \"is_wide\": true\n  },\n  \"name\": {\n    \"is_wide\": true,\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"object_name\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"filter_scope\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"columns\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"filter_fields\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"scrolling_mode\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"sort\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"show_count\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"type\": {\n    \"amis\": {\n      \"hidden\": true\n    }\n  },\n  \"shared\":{\n    \"amis\":{\n      \"visibleOn\":\"${global.user.is_space_admin}\"\n  }\n}\n}",
+                      "fields": [
+                        "label",
+                        "name",
+                        "object_name",
+                        "filter_scope",
+                        "show_count",
+                        "columns.$.field",
+                        "columns.$.width",
+                        "sort.$.field_name",
+                        "sort.$.order",
+                        "filters",
+                        "mobile_columns.$.field",
+                        "searchable_fields.$.field",
+                        "is_system",
+                        "shared"
+                      ],
                       "onEvent": {
                         "submitSucc": {
                           "weight": 0,
@@ -62,82 +78,4 @@ export const copyListview = ()=>{
           }
         }
     }
-}
-
-function fields(){
-  return [
-    "label",
-    "name",
-    "object_name",
-    "filter_scope",
-    "show_count",
-    "columns.$.field",
-    "columns.$.width",
-    "sort.$.field_name",
-    "sort.$.order",
-    "filters",
-    "mobile_columns.$.field",
-    "searchable_fields.$.field",
-    "is_system",
-    "shared"
-  ]
-}
-
-function fieldsExtend(){
-  return {
-    "label": {
-      "is_wide": true
-    },
-    "name": {
-      "is_wide": true,
-      "amis": {
-        "hidden": true
-      }
-    },
-    "object_name": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "filter_scope": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "columns": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "filter_fields": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "scrolling_mode": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "sort": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "show_count": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "type": {
-      "amis": {
-        "hidden": true
-      }
-    },
-    "shared": {
-      "amis": {
-        "visibleOn": "${global.user.is_space_admin}"
-      }
-    }
-  }
 }
