+++ packages/@steedos-widgets/amis-lib/src/lib/converter/amis/toolbars/edit/deleteListview.js
@@ -1,4 +1,5 @@
 export const deleteListview = ()=>{
+  
     return {
         "type": "button",
         "label": "删除",
@@ -24,8 +25,8 @@ export const deleteListview = ()=>{
                     "messages": {
                       "success": "删除成功"
                     },
-                    "requestAdaptor": requestAdaptor(),
-                    "adaptor": adaptor(),
+                    "requestAdaptor": "const { recordId } = api.body;\nvar deleteArray = [];\nif (recordId) { deleteArray.push(`delete:object_listviews__delete(id: \"${recordId}\")`); }\napi.data = { query: `mutation{${deleteArray.join(',')}}` };\n  return api;\n",
+                    "adaptor": "if (payload.errors) {\n  payload.status = 2;\n  payload.msg = payload.errors[0].message;\n}\nreturn payload;",
                   }
                 }
               },
@@ -41,25 +42,4 @@ export const deleteListview = ()=>{
           }
         }
     }
-}
-
-
-function requestAdaptor(){
-  return `
-    const { recordId } = api.body;
-    var deleteArray = [];
-    if (recordId) { deleteArray.push(\`delete:object_listviews__delete(id: "\${recordId}")\`); }
-    api.data = { query: \`mutation{\${deleteArray.join(',')}}\` };
-    return api;
-  `
-}
-
-function adaptor(){
-  return  `
-    if (payload.errors) {
-      payload.status = 2;
-      payload.msg = payload.errors[0].message;
-    }
-    return payload;
-  `
 }
