+++ packages/@steedos-widgets/amis-lib/src/lib/converter/amis/toolbars/edit/renameListview.js
@@ -1,4 +1,5 @@
 export const renameListview = ()=>{
+  
     return {
         "type": "button",
         "label": "重命名",
